requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
langchain>=0.2
langchain-openai>=0.1.20
pydantic>=2.0
python-dotenv>=1.0
faster-whisper>=1.0
soundfile>=0.12
